refactor(Views): create internal module for teachers
