+++ api/serializers/teacher/dashboard.py
@@ -5,36 +5,13 @@ from api.models.auxil.constants import (
     TEACHER_PEER_SUPPORT_FIELD_NAME_PREFIX,
     TEACHER_PEER_SUPPORT_OPTIONS,
 )
-from api.serializers import (
-    AgeRangeSerializer,
-    DashboardPersonalInfoSerializer,
-    DayAndTimeSlotSerializer,
-    LanguageAndLevelSerializer,
-    NonTeachingHelpSerializer,
-)
+from api.serializers import DashboardPersonalInfoSerializer
 from api.serializers.age_range import AgeRangeStringField
 from api.serializers.day_and_time_slot import MinifiedDayAndTimeSlotSerializer
 from api.serializers.group.minified import MinifiedGroupSerializer
 from api.serializers.language_and_level import MinifiedLanguageAndLevelSerializer
 from api.serializers.non_teaching_help import NonTeachingHelpSerializerField
 from api.serializers.utc_timedelta import UTCTimedeltaField
-
-
-class TeacherWriteSerializer(serializers.ModelSerializer[Teacher]):
-    class Meta:
-        model = Teacher
-        fields = "__all__"
-
-
-class TeacherReadSerializer(serializers.ModelSerializer[Teacher]):
-    student_age_ranges = AgeRangeSerializer(many=True, read_only=True)
-    teaching_languages_and_levels = LanguageAndLevelSerializer(many=True, read_only=True)
-    availability_slots = DayAndTimeSlotSerializer(many=True, read_only=True)
-    non_teaching_help_provided = NonTeachingHelpSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Teacher
-        fields = "__all__"
 
 
 class PeerSupportField(serializers.SerializerMethodField):
